+++ packages/ui/voo_data_grid/lib/src/presentation/widgets/voo_data_grid_stateless.dart
@@ -203,6 +203,7 @@ class VooDataGridStateless<T> extends StatelessWidget {
       onPrimaryFilterChanged: onPrimaryFilterChanged,
       showPrimaryFilters: showPrimaryFilters,
       combineFiltersAndPrimaryFilters: combineFiltersAndPrimaryFilters,
+      onRefresh: onRefresh,
     );
   }
 }